5000: refactored signals flow, changed colors to base(from figma/kit)
