+++ src/ducks/Signals/GQL/SignalsGQL.js
