+++ src/ducks/Signals/Redux/actions.spec.js
