+++ src/ducks/Signals/Redux/reducers.js
