+++ src/ducks/Signals/SignalFormManager/AboutForm/AboutForm.js
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types'
 
 import { Formik, Form } from 'formik'
 import { Button } from '@santiment-network/ui'
-import FormikInput from './../../components/formik-santiment-ui/FormikInput'
-import styles from './TriggerForm.module.scss'
+import FormikInput from '../../../../components/formik-santiment-ui/FormikInput'
+import styles from '../SignalCrudForm/Signal/TriggerForm.module.scss'
 
 const AboutForm = ({ triggerMeta, isEdit = false, onSubmit, onBack }) => {
   const [trigger, setTrigger] = useState(triggerMeta)
