+++ src/ducks/Signals/SignalFormManager/AboutForm/AboutForm.spec.js
