+++ src/ducks/Signals/SignalFormManager/SignalCrudForm/Signal/TriggerForm.spec.js
