+++ src/ducks/Signals/SignalModal/SignalMasterModalForm.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { connect } from 'react-redux'
 import { Modal, Button, Icon } from '@santiment-network/ui'
-import SignalMaster from './SignalMaster'
-import { checkIsLoggedIn } from './../../pages/UserSelectors'
+import SignalMaster from '../SignalFormManager/SignalMaster'
+import { checkIsLoggedIn } from '../../../pages/UserSelectors'
 import styles from './SignalMasterModalForm.module.scss'
 
 const SignalMasterModalForm = ({
