+++ src/ducks/Signals/Utils/utils.js
