+++ src/ducks/Signals/Utils/utils.spec.js
