+++ src/variables.scss
@@ -10,6 +10,3 @@ $disabled: #d3d3d3;
 $input-color-night: #a0a0a0;
 $input-placeholder-color-day: #999;
 $input-placeholder-color-night: #4b4b4b;
-
-$trigger-border: #e7eaf3;
-$trigger-header-bg: #f9fafc;
