+++ stories/SignalMasterStory.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { Provider } from 'react-redux'
 import { MockedProvider } from 'react-apollo/test-utils'
-import { SignalMaster } from './../src/ducks/Signals/SignalMaster'
+import { SignalMaster } from '../src/ducks/Signals/SignalFormManager/SignalMaster'
 import { allProjectsForSearchGQL } from './../src/pages/Projects/allProjectsGQL'
 import store from './store'
